# server/docs/swagger.yaml
definitions:
  common.Attachment:
    properties:
      name:
        type: string
      url:
        type: string
    type: object
  common.AttachmentMap:
    additionalProperties:
      items:
        $ref: '#/definitions/common.Attachment'
      type: array
    type: object
  common.JSONMap:
    additionalProperties: true
    type: object
...
  house.Resource:
    properties:
      ID:
        description: 主键ID
        type: integer
      approval_status:
        description: 审批状态
        type: string
      area:
        description: 房源面积
        type: string
      attachments:
        allOf:
        - $ref: '#/definitions/common.AttachmentMap'
        description: 房源图片
      building_id:
        description: 楼栋ID
        type: string
      city:
        description: 所属城市
        type: string
      click:
        description: 电话获取次数
        type: integer
      commission_price:
        description: 返佣金额
        type: integer
      createdAt:
        description: 创建时间
        type: string
      district_ids:
        description: 所属商圈ID串
        type: string
      districts:
        description: 所属商圈
        type: string
      door_no:
        description: 门牌号
        type: string
      feature:
        description: 房源特色
        type: string
      floor:
        description: 楼层
        type: string
      follow:
        description: 关注次数
        type: integer
      hasPic:
        description: 是否有照片
        type: boolean
      house_id:
        description: 房号ID
        type: string
      house_type:
        description: 房屋类型
        type: string
      is_team_house:
        description: 是否团队房源
        type: boolean
      owner:
        description: 房源归属用户ID
        type: integer
      phone:
        description: 联系手机号
        type: string
      price:
        description: 房源价格
        type: integer
      region:
        description: 行政区
        type: string
      remarks:
        description: 备注信息
        type: string
      rent_type:
        description: 出租类型
        type: string
      room_code:
        description: 房间号
        type: string
      room_number:
        description: 房间数量
        type: integer
      shared:
        description: 分享次数
        type: integer
      status:
        description: 状态
        type: string
      unit_id:
        description: 单元ID
        type: string
      updated_last_at:
        description: 最后编辑时间
        type: string
      updatedAt:
        description: 更新时间
        type: string
      view:
        description: 浏览次数
        type: integer
      xiaoqu:
        description: 所属小区名称
        type: string
      xiaoqu_id:
        description: 所属小区ID
        type: integer
    type: object
...
  request.ContactQuotaCreate:
    properties:
      amount:
        description: 增加次数
        type: integer
      remark:
        description: 备注
        type: string
      userPhone:
        description: 经纪人手机号
        type: string
    type: object
  request.ContactQuotaSearch:
    properties:
      keyword:
        description: 关键字
        type: string
      page:
        description: 页码
        type: integer
      pageSize:
        description: 每页大小
        type: integer
      userPhone:
        description: 经纪人手机号
        type: string
    type: object
...
  request.ResourceShareCreate:
    properties:
      expireDays:
        description: 有效天数
        type: integer
    type: object
  request.RewardApplicationAction:
    properties:
      action:
        description: 操作类型
        type: string
      id:
        description: 申请单ID
        type: integer
    type: object
  request.RewardApplicationCreate:
    properties:
      remark:
        description: 申请备注
        type: string
      resourceId:
        description: 房源ID
        type: integer
    type: object
  request.RewardApplicationSearch:
    properties:
      auditStatus:
        description: 后台审核状态
        type: string
      keyword:
        description: 关键字
        type: string
      page:
        description: 页码
        type: integer
      pageSize:
        description: 每页大小
        type: integer
      publisherConfirmStatus:
        description: 发布人确认状态
        type: string
    type: object
...
  response.PageResult:
    properties:
      list: {}
      page:
        type: integer
      pageSize:
        type: integer
      total:
        description: 总数
        type: integer
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
  response.RewardApplicationResponse:
    properties:
      ID:
        description: 主键ID
        type: integer
      applyUserId:
        description: 申请人ID
        type: integer
      applyUserPhone:
        description: |-
          下面这些联系人信息是申请创建时的快照，
          避免用户后续修改手机号/微信号后，历史审核单展示被“穿透更新”。
        type: string
      applyUserWxNo:
        description: 申请人微信号快照
        type: string
      auditStatus:
        description: 后台审核状态
        type: string
      createdAt:
        description: 创建时间
        type: string
      doorNo:
        description: 户室号
        type: string
      lastOperatedAtUnixMilli:
        description: 最后操作时间毫秒
        type: integer
      publisherConfirmStatus:
        description: 发布人确认状态
        type: string
      publisherUserId:
        description: 发布人ID
        type: integer
      publisherUserPhone:
        description: 发布人手机号快照
        type: string
      publisherUserWxNo:
        description: 发布人微信号快照
        type: string
      remark:
        description: 申请备注
        type: string
      resourceId:
        description: 房源ID
        type: integer
      updatedAt:
        description: 更新时间
        type: string
      xiaoqu:
        description: 小区名称
        type: string
    type: object
  response.SharedMapResponse:
    properties:
      expireAt:
        description: 分享过期时间
        type: integer
      list:
        description: 小区聚合点位
        items:
          $ref: '#/definitions/response.SharedMapXiaoqu'
        type: array
      userInfo:
        allOf:
        - $ref: '#/definitions/system.SysUser'
        description: 发布人资料
    type: object
  response.SharedMapXiaoqu:
    properties:
      latitude:
        description: 坐标纬度
        type: string
      longitude:
        description: 坐标经度
        type: string
      name:
        description: 小区名称
        type: string
      num:
        description: 房源数量
        type: integer
      xiaoquId:
        description: 小区ID
        type: integer
    type: object
...
  system.Meta:
    properties:
      activeName:
        type: string
      closeTab:
        description: 自动关闭tab
        type: boolean
      defaultMenu:
        description: 是否是基础路由（开发中）
        type: boolean
      icon:
        description: 菜单图标
        type: string
      keepAlive:
        description: 是否缓存
        type: boolean
      title:
        description: 菜单名
        type: string
      transitionType:
        description: 路由切换动画
        type: string
    type: object
...
  system.SysAuthority:
    properties:
      authorityId:
        description: 角色ID
        type: integer
      authorityName:
        description: 角色名
        type: string
      children:
        items:
          $ref: '#/definitions/system.SysAuthority'
        type: array
      createdAt:
        description: 创建时间
        type: string
      dataAuthorityId:
        items:
          $ref: '#/definitions/system.SysAuthority'
        type: array
      defaultRouter:
        description: 默认菜单(默认dashboard)
        type: string
      deletedAt:
        type: string
      menus:
        items:
          $ref: '#/definitions/system.SysBaseMenu'
        type: array
      parentId:
        description: 父角色ID
        type: integer
      updatedAt:
        description: 更新时间
        type: string
    type: object
  system.SysBaseMenu:
    properties:
      ID:
        description: 主键ID
        type: integer
      authoritys:
        items:
          $ref: '#/definitions/system.SysAuthority'
        type: array
      children:
        items:
          $ref: '#/definitions/system.SysBaseMenu'
        type: array
      component:
        description: 对应前端文件路径
        type: string
      createdAt:
        description: 创建时间
        type: string
      hidden:
        description: 是否在列表隐藏
        type: boolean
      menuBtn:
        items:
          $ref: '#/definitions/system.SysBaseMenuBtn'
        type: array
      meta:
        allOf:
        - $ref: '#/definitions/system.Meta'
        description: 附加属性
      name:
        description: 路由name
        type: string
      parameters:
        items:
          $ref: '#/definitions/system.SysBaseMenuParameter'
        type: array
      parentId:
        description: 父菜单ID
        type: integer
      path:
        description: 路由path
        type: string
      sort:
        description: 排序标记
        type: integer
      updatedAt:
        description: 更新时间
        type: string
    type: object
  system.SysBaseMenuBtn:
    properties:
      ID:
        description: 主键ID
        type: integer
      createdAt:
        description: 创建时间
        type: string
      desc:
        type: string
      name:
        type: string
      sysBaseMenuID:
        type: integer
      updatedAt:
        description: 更新时间
        type: string
    type: object
  system.SysBaseMenuParameter:
    properties:
      ID:
        description: 主键ID
        type: integer
      createdAt:
        description: 创建时间
        type: string
      key:
        description: 地址栏携带参数的key
        type: string
      sysBaseMenuID:
        type: integer
      type:
        description: 地址栏携带参数为params还是query
        type: string
      updatedAt:
        description: 更新时间
        type: string
      value:
        description: 地址栏携带参数的值
        type: string
    type: object
...
  system.SysUser:
    properties:
      ID:
        description: 主键ID
        type: integer
      authorities:
        description: 多用户角色
        items:
          $ref: '#/definitions/system.SysAuthority'
        type: array
      authority:
        allOf:
        - $ref: '#/definitions/system.SysAuthority'
        description: 用户角色
      authorityId:
        description: 用户角色ID
        type: integer
      contactViewQuotaTotal:
        description: 该次数用于房东房源联系方式查看扣减。
        type: integer
      createdAt:
        description: 创建时间
        type: string
      email:
        description: 用户邮箱
        type: string
      enable:
        description: 是否冻结
        type: integer
      headerImg:
        description: 头像
        type: string
      isFindHouseSupermarket:
        description: |-
          找房超市标识决定两件事：
          1. 该用户发布的房源会被标记为团队房源
          2. 该用户可以看到团队房源筛选和团队房源数据
        type: boolean
      nickName:
        description: 系统昵称
        type: string
      openid:
        description: openid
        type: string
      originSetting:
        allOf:
        - $ref: '#/definitions/common.JSONMap'
        description: 配置
      phone:
        description: 用户手机号
        type: string
      publishQuotaTotal:
        description: 上架额度限制的是“同时处于待出租状态的房源数量”。
        type: integer
      updatedAt:
        description: 更新时间
        type: string
      userName:
        description: 用户登录名
        type: string
      uuid:
        description: 用户UUID
        type: string
      wxNickName:
        description: 微信昵称
        type: string
      wxNo:
        description: 微信号
        type: string
    type: object
...
info:
  contact: {}
  description: 使用gin+vue进行极速开发的全栈开发基础平台
  title: Gin-Vue-Admin Swagger API接口文档
  version: v2.8.4
paths:
...
  /api/house/contactQuota/grant:
    post:
      consumes:
      - application/json
      description: '[新增接口] 按经纪人手机号给账号充值房东房源联系方式查看次数。'
      parameters:
      - description: 增加次数参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.ContactQuotaCreate'
      produces:
      - application/json
      responses:
        "200":
          description: 操作结果
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  type: string
                msg:
                  type: string
              type: object
      summary: '[新增] 增加经纪人联系方式查看次数'
      tags:
      - Admin
  /api/house/contactQuota/list:
    post:
      consumes:
      - application/json
      description: '[新增接口] 返回充值和消耗流水，便于后台核对剩余次数和使用情况。'
      parameters:
      - description: 查询参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.ContactQuotaSearch'
      produces:
      - application/json
      responses:
        "200":
          description: 次数流水
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.PageResult'
                msg:
                  type: string
              type: object
      summary: '[新增] 查看联系方式次数流水'
      tags:
      - Admin
...
  /api/house/reward/action:
    post:
      consumes:
      - application/json
      description: '[新增接口] 支持审核中、审通过待发放、已发放、未通过等后台状态流转。'
      parameters:
      - description: 操作参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.RewardApplicationAction'
      produces:
      - application/json
      responses:
        "200":
          description: 操作结果
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  type: string
                msg:
                  type: string
              type: object
      summary: '[新增] 后台操作成交有礼审核状态'
      tags:
      - Admin
  /api/house/reward/list:
    post:
      consumes:
      - application/json
      description: '[新增接口] 后台只展示“发布人已确认”的成交有礼申请单。'
      parameters:
      - description: 查询参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.RewardApplicationSearch'
      produces:
      - application/json
      responses:
        "200":
          description: 审核列表
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.PageResult'
                msg:
                  type: string
              type: object
      summary: '[新增] 后台成交有礼审核列表'
      tags:
      - Admin
...
  /center/house/share:
    get:
      description: '[新增接口] 未登录也可访问；先返回按小区聚合后的地图点位，前端点击点位后再调用 /center/house/share/list
        查看该小区下的房源。'
      parameters:
      - description: 分享token
        in: query
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 分享房源地图点位
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.SharedMapResponse'
                msg:
                  type: string
              type: object
      summary: '[新增] 通过分享token获取地图点位列表'
      tags:
      - Center
    post:
      consumes:
      - application/json
      description: '[新增接口] 为当前登录用户生成 7 天有效的房源地图分享 token。'
      parameters:
      - description: 分享参数
        in: body
        name: data
        schema:
          $ref: '#/definitions/request.ResourceShareCreate'
      produces:
      - application/json
      responses:
        "200":
          description: 分享结果
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  additionalProperties: true
                  type: object
                msg:
                  type: string
              type: object
      summary: '[新增] 创建我的房源地图分享链接'
      tags:
      - Center
  /center/house/share/list:
    get:
      description: '[新增接口] 地图点位点击后调用，仍然通过 ZincSearch 按 owner + xiaoqu_id 过滤，再回表补全房源详情。'
      parameters:
      - description: 分享token
        in: query
        name: token
        required: true
        type: string
      - description: 小区ID
        in: query
        name: xiaoquId
        required: true
        type: integer
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 分享小区房源列表
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/response.PageResult'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/house.Resource'
                        type: array
                    type: object
                msg:
                  type: string
              type: object
      summary: '[新增] 通过分享token和小区ID获取点位下房源列表'
      tags:
      - Center
...
  /center/reward/apply:
    post:
      consumes:
      - application/json
      description: '[新增接口] 申请人提交房源和备注后，后端会自动补齐申请人/发布人的手机号和微信号快照。'
      parameters:
      - description: 申请参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.RewardApplicationCreate'
      produces:
      - application/json
      responses:
        "200":
          description: 申请结果
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  type: string
                msg:
                  type: string
              type: object
      summary: '[新增] 发起出房有礼申请'
      tags:
      - Center
  /center/reward/publisher/action:
    post:
      consumes:
      - application/json
      description: '[新增接口] 发布人确认后，申请单才会进入后台审核状态。'
      parameters:
      - description: 操作参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.RewardApplicationAction'
      produces:
      - application/json
      responses:
        "200":
          description: 操作结果
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  type: string
                msg:
                  type: string
              type: object
      summary: '[新增] 发布人确认或拒绝出房有礼申请'
      tags:
      - Center
  /center/reward/publisher/list:
    post:
      consumes:
      - application/json
      description: '[新增接口] 房源发布人查看自己的申请单列表，不等同于后台总审核池。'
      parameters:
      - description: 查询参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.RewardApplicationSearch'
      produces:
      - application/json
      responses:
        "200":
          description: 审核列表
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/response.PageResult'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/response.RewardApplicationResponse'
                        type: array
                    type: object
                msg:
                  type: string
              type: object
      summary: '[新增] 发布人查看出房有礼审核列表'
      tags:
      - Center
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"
...